pyproject.toml
[project]
name = "faster-rcnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.0",
    "torch>=2.9.1",
    "torchinfo>=1.8.0",
    "torchvision>=0.24.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
